Implementation of proper database, users can now be created, recycler view is now fixed and makes logging more interactive.
user data does NOT save when you wipe the phone data (and change the version in database) so that should be implemented as well.

Strotta.java may need work as one set of parameters are ignored (to allow the program to run at the moment)

removed some irrelevant files that were deprecated and were causing problems.